event_bot_mvp/requirements.txt
apscheduler
python-telegram-bot
SQLAlchemy
langchain_ollama
langchain
